[Refactoring ⚙️] 회원가입 Profile 타입스크립트 무시 제거
