+++ src/components/Input/components/InputInput.tsx
@@ -1,12 +1,15 @@
-// @ts-nocheck
-import IconButton from "@components/IconButton/IconButton.js"
+import { InputHTMLAttributes } from "react"
 
 import * as S from "../StyledInput.js"
 
-const InputInput = ({ type, props, ...rest }) => {
+interface InputInputProps extends InputHTMLAttributes<HTMLInputElement> {
+  // any 타입 고치기
+  props: any
+}
+
+const InputInput = ({ props, ...rest }: InputInputProps) => {
   return (
     <S.Input
-      type={type}
       {...props}
       {...rest}
     />
