--- src/pages/Signup/Profile/StyledProfile.ts
@@ -1,8 +1,0 @@
-import styled from "styled-components"
-
-export const ProfileFormWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  width: 100%;
-`
